实现crc校验

+++ Bsp/Uart/Inc/uart.h
@@ -2,7 +2,7 @@
  * @Author: Hengyang Jiang
  * @Date: 2024-12-12 21:34:06
  * @LastEditors: Hengyang Jiang
- * @LastEditTime: 2024-12-13 15:04:10
+ * @LastEditTime: 2024-12-16 13:38:30
  * @Description: uart.h
  *
  * Copyright (c) 2024 by https://github.com/Nolan-Jon, All Rights Reserved.
@@ -25,12 +25,12 @@ typedef struct
     /* data */
     uint8_t idx;                                   /* 串口实例的编号,对于该开发板来讲,编号只有1和3 */
     uint8_t recv_buffer[UART_RECEIVE_BUFFER_SIZE]; /* 接收缓冲区 */
-    uint8_t recv_buffer_size;                      /* 接收一包数据的大小 */
+    uint16_t recv_buffer_size;                     /* 接收一包数据的大小:数据帧大小 */
     UART_HandleTypeDef *uartHandle;                /* 每个UART实例对应的设备句柄 */
     uart_recv_decode uart_recv_decode_callback;    /* 解析协议回调函数 */
 } UART_InstanceDef;
 typedef UART_InstanceDef *UART_InstanceHandle;
 
-UART_InstanceHandle Y_uart_create_instance(uint8_t, uint8_t, UART_HandleTypeDef *, uart_recv_decode);
+UART_InstanceHandle Y_uart_create_instance(uint8_t, uint16_t, UART_HandleTypeDef *, uart_recv_decode);
 
 #endif //!__UART__H__
